Correção da lista de contas ao cancelar o edit()

+++ src/pages/contas/contas.ts
@@ -1,6 +1,7 @@
 import { Component } from '@angular/core';
+import { Toast } from 'ionic-native';
 
-import { ModalController } from 'ionic-angular';
+import { ModalController, AlertController, NavController} from 'ionic-angular';
 import { DAOContas } from '../../app/dao/dao-contas';
 import { ModalContasPage} from '../modal-contas/modal-contas';
 
@@ -12,9 +13,17 @@ export class ContasPage {
   dao: any;
   listContas: any;
   modalCtrl: any;
+  alertCtrl: any;
+  navCtrl: any;
 
-  constructor(modalCtrl: ModalController) {
+  constructor(
+    modalCtrl: ModalController,
+    alertCtrl: AlertController,
+    navCtrl: NavController
+  ) {
     this.modalCtrl = modalCtrl;
+    this.alertCtrl = alertCtrl;
+    this.navCtrl = navCtrl;
     this.dao = new DAOContas;
     this.dao.getList((lista) => {
       this.listContas = lista;
@@ -24,9 +33,15 @@ export class ContasPage {
   insert() {
     let modal = this.modalCtrl.create(ModalContasPage);
     modal.onDidDismiss(data => {
-      this.dao.insert(data, (data) => {
-        this.listContas.push(data);
-      });
+      if (data) {
+        this.dao.insert(data, (data) => {
+          this.listContas.push(data);
+          Toast.showShortBottom("Conta Inserida Com Sucesso !").subscribe(
+            toast => {
+              console.log(toast);
+            });
+        });
+      }
     });
     modal.present();
   }
@@ -34,17 +49,50 @@ export class ContasPage {
   edit(conta) {
     let modal = this.modalCtrl.create(ModalContasPage, {parametro: conta});
     modal.onDidDismiss(data => {
-      this.dao.edit(data, (data) => {
-      });
+      if (data) {
+        this.dao.edit(data, (data) => {
+          Toast.showShortBottom("Conta Alterada Com Sucesso !").subscribe(
+            toast => {
+              console.log(toast);
+            });
+        });
+      } else {
+        this.dao.getList((lista) => {
+          this.listContas = lista;
+        });
+      }
     });
     modal.present();
   }
 
   delete(conta) {
-    this.dao.delete(conta, (data) => {
-      let pos = this.listContas.indexOf(data);
-      this.listContas.splice(pos, 1);
+    let confirm = new this.alertCtrl.create({
+      title: 'Excluir Conta',
+      message: "Gostaria realmente de excluir a conta: " + conta.descricao + "?",
+      buttons: [
+        {
+          text: 'Sim',
+          handler: () => {
+            this.dao.delete(conta, (data) => {
+              let pos = this.listContas.indexOf(data);
+              this.listContas.splice(pos, 1);
+              Toast.showShortBottom("Conta Excluída Com Sucesso !").subscribe(
+                toast => {
+                  console.log(toast);
+                });
+            });
+          }
+        },
+        {
+          text: 'Não',
+          role: 'cancel',
+          handler: () => {
+            console.log('Exclusão Cancelada');
+          }
+        },
+      ]
     });
+    this.navCtrl.push(confirm);
   }
 
 }
